Remove BeforeAfterHookMixin and run the hooks separately
Turns out that the mixin is not always applicable as there might be instances where there are more than one 'before'/'after' hooks that get run in the same view
